Extract PoolerConfig from config.py
